Refactor fetchPanoptesUsers as utils function
