+++ packages/lib-user/src/index.js
@@ -38,6 +38,7 @@ export { createPanoptesMembership } from './utils/createPanoptesMembership.js'
 export { createPanoptesUserGroup } from './utils/createPanoptesUserGroup.js'
 export { deletePanoptesMembership } from './utils/deletePanoptesMembership.js'
 export { deletePanoptesUserGroup } from './utils/deletePanoptesUserGroup.js'
+export { fetchPanoptesUsers } from './utils/fetchPanoptesUsers.js'
 export { getDateInterval } from './utils/getDateInterval.js'
 export { updatePanoptesMembership } from './utils/updatePanoptesMembership.js'
 export { updatePanoptesUserGroup } from './utils/updatePanoptesUserGroup.js'
