+++ packages/lib-user/src/utils/index.js
@@ -3,6 +3,7 @@ export { createPanoptesMembership } from './createPanoptesMembership.js'
 export { createPanoptesUserGroup } from './createPanoptesUserGroup.js'
 export { deletePanoptesMembership } from './deletePanoptesMembership.js'
 export { deletePanoptesUserGroup } from './deletePanoptesUserGroup.js'
+export { fetchPanoptesUsers } from './fetchPanoptesUsers.js'
 export { getDateInterval } from './getDateInterval.js'
 export { updatePanoptesMembership } from './updatePanoptesMembership.js'
 export { updatePanoptesUserGroup } from './updatePanoptesUserGroup.js'
